#174 renamed SettingValueAccessor to SettingDataAccessor
