--- src/libs/ff1-coredefs/dva_std_collections.h
@@ -1,16 +1,0 @@
-#pragma once
-
-#if 0
-//#include "FFHValue.h"
-struct FFHValue;
-#include <vector>
-
-namespace ffh
-{
-	namespace fda
-	{
-		const FFHValue& operator>>(const FFHValue& stg, std::vector<std::string>& value);
-		FFHValue& operator<<(FFHValue& stg, const std::vector<std::string>& value);
-	}
-}
-#endif
