+++ src/libs/ff1-coredefs/ff1-coredefs.vcxproj
@@ -190,8 +190,6 @@
     <ClCompile Include="dialogue_helpers.cpp" />
     <ClCompile Include="DirPathRemover.cpp" />
     <ClCompile Include="draw_functions.cpp" />
-    <ClCompile Include="dva_primitives.cpp" />
-    <ClCompile Include="dva_std_collections.cpp" />
     <ClCompile Include="Editors2.cpp" />
     <ClCompile Include="FFH2Project.cpp" />
     <ClCompile Include="FFHacksterProject.cpp" />
@@ -213,7 +211,7 @@
     <ClCompile Include="ResourceHandleScope.cpp" />
     <ClCompile Include="EditorSettingsSerializer.cpp" />
     <ClCompile Include="SetResourceHandleScope.cpp" />
-    <ClCompile Include="SettingValueAccessor.cpp" />
+    <ClCompile Include="SettingDataAccessor.cpp" />
     <ClCompile Include="set_types.cpp" />
     <ClCompile Include="stdafx.cpp">
       <PrecompiledHeader Condition="'$(Configuration)|$(Platform)'=='Debug|Win32'">Create</PrecompiledHeader>
@@ -224,7 +222,6 @@
     <ClCompile Include="std_assembly.cpp" />
     <ClCompile Include="stream_ops.cpp" />
     <ClCompile Include="string_functions.cpp" />
-    <ClCompile Include="sva_primitives.cpp" />
     <ClCompile Include="type_support.cpp" />
     <ClCompile Include="Upgrades_ini.cpp" />
     <ClCompile Include="upgrade_functions.cpp" />
@@ -262,10 +259,8 @@
     <ClInclude Include="DATA.h" />
     <ClInclude Include="datanode.h" />
     <ClInclude Include="ValueDataAccessor.h" />
-    <ClInclude Include="dva_primitives.h" />
-    <ClInclude Include="dva_std_collections.h" />
     <ClInclude Include="FFHImages.h" />
-    <ClInclude Include="SettingValueAccessor.h" />
+    <ClInclude Include="SettingDataAccessor.h" />
     <ClInclude Include="dialogue_helpers.h" />
     <ClInclude Include="DirPathRemover.h" />
     <ClInclude Include="draw_functions.h" />
@@ -314,7 +309,6 @@
     <ClInclude Include="stream_ops.h" />
     <ClInclude Include="string_conversions.hpp" />
     <ClInclude Include="string_functions.h" />
-    <ClInclude Include="sva_primitives.h" />
     <ClInclude Include="targetver.h" />
     <ClInclude Include="type_support.h" />
     <ClInclude Include="Upgrades_ini.h" />
