+++ src/libs/ff1-editors/Attack.cpp
@@ -5,8 +5,6 @@
 #include "Attack.h"
 #include "AsmFiles.h"
 #include <core_exceptions.h>
-#include <ValueDataAccessor.h>
-#include <dva_primitives.h>
 #include "editor_label_functions.h"
 #include <FFH2Project.h>
 #include "GameSerializer.h"
@@ -17,6 +15,8 @@
 #include "io_functions.h"
 #include "string_functions.h"
 #include "ui_helpers.h"
+#include <ValueDataAccessor.h>
+#include <vda_std_collections.h>
 
 using namespace Editorlabels;
 using namespace Imaging;
