+++ src/libs/ff1-editors/Text.cpp
@@ -4,21 +4,22 @@
 #include "stdafx.h"
 #include "Text.h"
 #include <FFH2Project.h>
-#include "collection_helpers.h"
-#include <core_exceptions.h>
-#include <ValueDataAccessor.h>
-#include <dva_primitives.h>
-#include <editor_label_functions.h>
-#include "general_functions.h"
-#include "ini_functions.h"
-#include "string_functions.h"
-#include "ingame_text_functions.h"
-#include "io_functions.h"
 #include "AsmFiles.h"
 #include "GameSerializer.h"
-#include "path_functions.h"
-#include "ui_helpers.h"
+#include "collection_helpers.h"
+#include <core_exceptions.h>
+#include <editor_label_functions.h>
+#include "general_functions.h"
 #include "imaging_helpers.h"
+#include "ingame_text_functions.h"
+#include "ini_functions.h"
+#include "io_functions.h"
+#include "path_functions.h"
+#include "string_functions.h"
+#include "ui_helpers.h"
+#include <ValueDataAccessor.h>
+#include <vda_std_collections.h>
+
 #include "NewLabel.h"
 #include <TextEditorSettings.h>
