+++ src/libs/ff1-settings/TextEditorSettings.cpp
@@ -3,7 +3,7 @@
 #include "FFHacksterProject.h"
 #include "ini_functions.h"
 #include "type_support.h"
-#include <SettingValueAccessor.h>
+#include <SettingDataAccessor.h>
 #include <sva_primitives.h>
 
 using namespace Ini;
@@ -52,7 +52,7 @@ void TextEditorSettings::SetDefaults()
 bool TextEditorSettings::Read()
 {
 	//READ_SETTING_BOOL(ShowActualText);
-	ffh::fda::SettingValueAccessor s(m_prj2, "text");
+	ffh::acc::SettingDataAccessor s(m_prj2, "text");
 	ShowActualText = s.get<bool>("ShowActualText");
 	return true;
 }
@@ -60,7 +60,7 @@ bool TextEditorSettings::Read()
 bool TextEditorSettings::Write()
 {
 	//WRITE_SETTING_BOOL(ShowActualText);
-	ffh::fda::SettingValueAccessor s(m_prj2, "text");
+	ffh::acc::SettingDataAccessor s(m_prj2, "text");
 	s.set("ShowActualText", ShowActualText);
 	return true;
 }
