+++ src/libs/ff1-utils/editor_label_functions.cpp
@@ -10,7 +10,7 @@
 #include <FFH2Project.h>
 #include <algorithm>
 #include <ValueDataAccessor.h>
-#include <dva_primitives.h>
+#include <vda_std_collections.h>
 
 using ffh::str::tomfc;
 using ffh::str::tostd;
